+++ seed/csharp-sdk/examples/no-custom-config/src/SeedExamples.Test/Unit/Serialization/MigrationTest.cs
@@ -1,0 +1,40 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using FluentAssertions.Json;
+using Newtonsoft.Json.Linq;
+using NUnit.Framework;
+using SeedExamples;
+
+#nullable enable
+
+namespace SeedExamples.Test;
+
+[TestFixture]
+public class MigrationTest
+{
+    [Test]
+    public void TestSerialization()
+    {
+        var inputJson =
+            @"
+        {
+          ""name"": ""001_init"",
+          ""status"": ""RUNNING""
+        }
+        ";
+
+        var serializerOptions = new JsonSerializerOptions
+        {
+            DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
+        };
+
+        var deserializedObject = JsonSerializer.Deserialize<Migration>(
+            inputJson,
+            serializerOptions
+        );
+
+        var serializedJson = JsonSerializer.Serialize(deserializedObject, serializerOptions);
+
+        JToken.Parse(inputJson).Should().BeEquivalentTo(JToken.Parse(serializedJson));
+    }
+}