+++ seed/csharp-sdk/examples/no-custom-config/src/SeedExamples.Test/Unit/Serialization/TreeTest.cs
@@ -1,0 +1,43 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using FluentAssertions.Json;
+using Newtonsoft.Json.Linq;
+using NUnit.Framework;
+using SeedExamples;
+
+#nullable enable
+
+namespace SeedExamples.Test;
+
+[TestFixture]
+public class TreeTest
+{
+    [Test]
+    public void TestSerialization()
+    {
+        var inputJson =
+            @"
+        {
+          ""nodes"": [
+            {
+              ""name"": ""left""
+            },
+            {
+              ""name"": ""right""
+            }
+          ]
+        }
+        ";
+
+        var serializerOptions = new JsonSerializerOptions
+        {
+            DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
+        };
+
+        var deserializedObject = JsonSerializer.Deserialize<Tree>(inputJson, serializerOptions);
+
+        var serializedJson = JsonSerializer.Serialize(deserializedObject, serializerOptions);
+
+        JToken.Parse(inputJson).Should().BeEquivalentTo(JToken.Parse(serializedJson));
+    }
+}