+++ seed/csharp-sdk/examples/no-custom-config/src/SeedExamples/Commons/CommonsClient.cs
@@ -1,0 +1,18 @@
+using SeedExamples.Core;
+
+#nullable enable
+
+namespace SeedExamples.Commons;
+
+public partial class CommonsClient
+{
+    private RawClient _client;
+
+    internal CommonsClient(RawClient client)
+    {
+        _client = client;
+        Types = new TypesClient(_client);
+    }
+
+    public TypesClient Types { get; }
+}