+++ seed/csharp-sdk/examples/no-custom-config/src/SeedExamples/Types/Types/Directory.cs
@@ -1,0 +1,23 @@
+using System.Text.Json.Serialization;
+using SeedExamples.Core;
+
+#nullable enable
+
+namespace SeedExamples;
+
+public record Directory
+{
+    [JsonPropertyName("name")]
+    public required string Name { get; set; }
+
+    [JsonPropertyName("files")]
+    public IEnumerable<File>? Files { get; set; }
+
+    [JsonPropertyName("directories")]
+    public IEnumerable<Directory>? Directories { get; set; }
+
+    public override string ToString()
+    {
+        return JsonUtils.Serialize(this);
+    }
+}