+++ seed/csharp-sdk/examples/no-custom-config/src/SeedExamples/Types/Types/StuntDouble.cs
@@ -1,0 +1,20 @@
+using System.Text.Json.Serialization;
+using SeedExamples.Core;
+
+#nullable enable
+
+namespace SeedExamples;
+
+public record StuntDouble
+{
+    [JsonPropertyName("name")]
+    public required string Name { get; set; }
+
+    [JsonPropertyName("actorOrActressId")]
+    public required string ActorOrActressId { get; set; }
+
+    public override string ToString()
+    {
+        return JsonUtils.Serialize(this);
+    }
+}