+++ generators/csharp/sdk/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Feature: Add support for generating `README.md` files.
+
 ## [1.5.0] - 2024-08-28
 
 - Feature: Support service-level headers.